Implement scene subscription.
Add ability in the client to show and hide

views created on the server dynamically through toolbar or view menu.

REveManager:

Add thread local variable to handle added and removed scene subscription.

Stream newly connected scenes in MIR execution thread.

Check the mandatory flag of the scene of the WindowConnect() function.

REveScene, REveViewer:

Add a mandatory flag to define what is going to be visible to the client on the first connection.

ui5:

Add a toolbar with controls in each view. Allow viewer destruction/creation on subscription changes.